refactor: change folder hierarchy
