+++ CMakeLists.txt
@@ -14,23 +14,27 @@ if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
 endif()
 
 find_package(ament_cmake REQUIRED)
-find_package(aruku REQUIRED)
 find_package(framework_common REQUIRED)
-find_package(kansei REQUIRED)
-find_package(ninshiki_opencv REQUIRED)
+find_package(keisan REQUIRED)
+find_package(OpenCV REQUIRED)
+find_package(rclcpp REQUIRED)
+find_package(tachimawari REQUIRED)
+find_package(tachimawari_interfaces REQUIRED)
 
 add_library(${PROJECT_NAME} SHARED
-  "src/${PROJECT_NAME}/head.cpp")
+  "src/${PROJECT_NAME}/head/head.cpp")
 
 target_include_directories(${PROJECT_NAME} PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>)
 
 ament_target_dependencies(${PROJECT_NAME}
-  aruku
   framework_common
-  kansei
-  ninshiki_opencv)
+  keisan
+  OpenCV
+  rclcpp
+  tachimawari
+  tachimawari_interfaces)
 
 install(DIRECTORY "include" DESTINATION ".")
 
@@ -40,11 +44,14 @@ install(TARGETS ${PROJECT_NAME}
   LIBRARY DESTINATION "lib"
   RUNTIME DESTINATION "bin")
 
-add_executable(main "src/head_main.cpp")
-target_link_libraries(main ${PROJECT_NAME})
+# add_executable(main "src/head_main.cpp")
+# target_include_directories(main PUBLIC
+#   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+#   $<INSTALL_INTERFACE:include>)
+# target_link_libraries(main ${PROJECT_NAME})
 
-install(TARGETS main
-  DESTINATION "lib/${PROJECT_NAME}")
+# install(TARGETS main
+#   DESTINATION "lib/${PROJECT_NAME}")
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
@@ -52,10 +59,12 @@ if(BUILD_TESTING)
 endif()
 
 ament_export_dependencies(
-  aruku
   framework_common
-  kansei
-  ninshiki_opencv)
+  keisan
+  OpenCV
+  rclcpp
+  tachimawari
+  tachimawari_interfaces)
 ament_export_include_directories("include")
 ament_export_libraries(${PROJECT_NAME})
 ament_package()
