+++ package.xml
@@ -7,10 +7,12 @@
   <maintainer email="jalilpro31@gmail.com">Maroqi Jalil</maintainer>
   <license>MIT License</license>
   <buildtool_depend>ament_cmake</buildtool_depend>
-  <depend>aruku</depend>
   <depend>framework_common</depend>
-  <depend>kansei</depend>
-  <depend>ninshiki_opencv</depend>
+  <depend>keisan</depend>
+  <depend>libopencv-dev</depend>
+  <depend>nlohmann-json3-dev</depend>
+  <depend>tachimawari</depend>
+  <depend>tachimawari_interfaces</depend>
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
   <export>
